add transformer.pause()
