+++ src/ramjet.js
@@ -1,7 +1,6 @@
 import { showNode, hideNode } from './utils/node';
 import wrapNode from './wrappers/wrapNode';
-import TimerTransformer from './transformers/TimerTransformer';
-import KeyframeTransformer from './transformers/KeyframeTransformer';
+import transformer from './transformers/transformer.js';
 import { linear, easeIn, easeOut, easeInOut } from './easing';
 import { keyframesSupported } from './utils/detect';
 
@@ -31,11 +30,7 @@ export default {
 		// stacking context than `to`. Not much we can do ¯\_(ツ)_/¯
 		to.setZIndex( Math.max( to.getZIndex(), from.getZIndex() + 1 ) );
 
-		if ( !keyframesSupported || options.useTimer ) {
-			return new TimerTransformer( from, to, options );
-		} else {
-			return new KeyframeTransformer( from, to, options );
-		}
+		return transformer( from, to, options );
 	},
 
 	hide ( ...nodes ) {
