+++ src/utils/detect.js
@@ -5,6 +5,7 @@ let TRANSFORM;
 let TRANSFORM_ORIGIN;
 let TRANSFORM_CSS;
 let KEYFRAMES;
+let ANIMATION;
 let ANIMATION_DIRECTION;
 let ANIMATION_DURATION;
 let ANIMATION_ITERATION_COUNT;
@@ -37,6 +38,7 @@ if (
 	if ( 'animation' in div.style ) {
 		KEYFRAMES = '@keyframes';
 
+		ANIMATION                 = 'animation';
 		ANIMATION_DIRECTION       = 'animationDirection';
 		ANIMATION_DURATION        = 'animationDuration';
 		ANIMATION_ITERATION_COUNT = 'animationIterationCount';
@@ -47,6 +49,7 @@ if (
 	} else {
 		KEYFRAMES = '@-webkit-keyframes';
 
+		ANIMATION                 = 'webkitAnimation';
 		ANIMATION_DIRECTION       = 'webkitAnimationDirection';
 		ANIMATION_DURATION        = 'webkitAnimationDuration';
 		ANIMATION_ITERATION_COUNT = 'webkitAnimationIterationCount';
@@ -65,6 +68,7 @@ export {
 	TRANSFORM_ORIGIN,
 	TRANSFORM_CSS,
 	KEYFRAMES,
+	ANIMATION,
 	ANIMATION_DIRECTION,
 	ANIMATION_DURATION,
 	ANIMATION_ITERATION_COUNT,
